OS X: compile util
This is a rollup of commits authored by Adar, JD, and me. These changes allow
the util module to be compiled on OS X.

+++ src/kudu/util/pb_util.cc
@@ -575,7 +575,7 @@ Status ReadablePBContainerFile::Close() {
 Status ReadablePBContainerFile::ValidateAndRead(size_t length, EofOK eofOK,
                                                 Slice* result, gscoped_ptr<uint8_t[]>* scratch) {
   // Validate the read length using the file size.
-  uint64_t file_size;
+  size_t file_size;
   RETURN_NOT_OK(reader_->Size(&file_size));
   if (offset_ + length > file_size) {
     switch (eofOK) {
